+++ src/main/java/ae2stuff/network/ModNetwork.java
@@ -1,0 +1,25 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2026 AE2 Stuff UD contributors
+ */
+
+package ae2stuff.network;
+
+import net.minecraftforge.fml.common.network.NetworkRegistry;
+import net.minecraftforge.fml.common.network.simpleimpl.SimpleNetworkWrapper;
+import net.minecraftforge.fml.relauncher.Side;
+
+import ae2stuff.Tags;
+
+public final class ModNetwork {
+
+    public static final SimpleNetworkWrapper CHANNEL = NetworkRegistry.INSTANCE.newSimpleChannel(Tags.MOD_ID);
+
+    private ModNetwork() {
+    }
+
+    public static void init() {
+        CHANNEL.registerMessage(PacketKitKey.Handler.class, PacketKitKey.class, 0, Side.SERVER);
+    }
+}